jrh-lisp: Add Org exporter to generate PowerPoint files.
This uses Kawa Scheme and Apache POI: the export generates a Kawa
script, which, when executed, uses the POI Java libraries to generate
the PowerPoint file from a provided prototype file.

The Kawa template here assumes a specific PowerPoint file to be used
as the skeleton.  Changes to that PowerPoint file (or, say, using an
actual PowerPoint template) will require changes to the Kawa template.