ox-pptx: Do a better job of fitting tables onto slides.
I think the key w.r.t. not hanging off the bottom was to
call (s:update-cell-anchor) after updating the column widths.  But
that starts them off all equally wide, so do a pass trying to adjust
widths to look for ones that shrink the overall table height.  This is
by no means optimal.

It still doesn't help if the table just has too much content.  I tried
setting autofit=normal and shrinking row heights, but that doesn't
seem to do anything.  It might be some kind of interaction with
the *table-style*?